requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
autogluon.timeseries>=1.0.0
plotly>=5.17.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
google-generativeai>=0.3.0
protobuf==3.20.3
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0